Refactor Incidents component: extract hero card and report sections into separate components, streamline state management, and remove unused code
